Adding x86 TLS

+++ src/gtirb-decoder/arch/X86Loader.cpp
@@ -135,12 +135,10 @@ std::optional<relations::Operand> X86Loader::build(const cs_x86_op& CsOp)
             return static_cast<int32_t>(CsOp.imm);
         case X86_OP_MEM:
         {
-            relations::IndirectOp I = {registerName(CsOp.mem.segment),
-                                       registerName(CsOp.mem.base),
-                                       registerName(CsOp.mem.index),
-                                       CsOp.mem.scale,
-                                       CsOp.mem.disp,
-                                       static_cast<uint64_t>(CsOp.size) * 8};
+            relations::IndirectOp I = {
+                registerName(CsOp.mem.segment),      registerName(CsOp.mem.base),
+                registerName(CsOp.mem.index),        CsOp.mem.scale,
+                static_cast<int32_t>(CsOp.mem.disp), static_cast<uint32_t>(CsOp.size) * 8};
             return I;
         }
         case X86_OP_INVALID:
